requirements.txt
python-telegram-bot
urllib3
google-cloud-storage
requests
flask
paho-mqtt==1.6.1
flask-mqtt